trabalho/app/requirements.txt
boto3==1.35.89
confluent-kafka==2.6.1
numpy==2.1.3
orjson==3.10.12
polars==1.35.0
pyarrow==18.1.0
pyiceberg[s3fs,sql-sqlite]==0.8.1
streamlit==1.40.2
pandas==2.2.3
plotly==5.24.1
psycopg2-binary==2.9.10
